Reorganize shared and linear math utilities
