+++ rocketsim/src/bullet/dynamics/constraint_solver/seq_impulse_constraint_solver.rs
@@ -6,10 +6,7 @@ use crate::bullet::{
         manifold_point::ManifoldPoint, persistent_manifold::PersistentManifold,
     },
     dynamics::rigid_body::{CollisionFlags, RigidBody},
-    linear_math::{
-        plane_space_1,
-        transform_util::{integrate_trans, integrate_trans_no_rot},
-    },
+    linear_math::{integrate_trans, integrate_trans_no_rot, plane_space_1},
 };
 
 struct SpecialResolveInfo {
