+++ rocketsim/src/shared/mod.rs
@@ -1,4 +1,5 @@
 mod aabb;
+mod angle;
 pub(crate) mod bvh;
 mod quad_ray;
 pub(crate) mod quantize;
@@ -6,4 +7,5 @@ pub mod rand;
 pub mod rsmath;
 
 pub use aabb::*;
+pub use angle::Angle;
 pub use quad_ray::*;
