+++ rocketsim/src/sim/ball/base.rs
@@ -18,10 +18,10 @@ use crate::{
             discrete_dynamics_world::DiscreteDynamicsWorld,
             rigid_body::{ActivationState, CollisionFlags, RigidBody, RigidBodyConstructionInfo},
         },
-        linear_math::angle::Angle,
     },
     consts::{BT_TO_UU, UU_TO_BT, dropshot, heatseeker, snowday},
     get_neighbor_indices_1, get_neighbor_indices_2, get_tile_pos,
+    shared::Angle,
     sim::{UserInfoTypes, consts},
 };
